DT-961 Move crn to be within otherIds on offender detail
