+++ src/test/kotlin/uk/gov/justice/digital/hmpps/indexer/health/HealthCheckIntegrationTest.kt
@@ -8,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.test.context.junit.jupiter.SpringExtension
 import org.springframework.test.util.ReflectionTestUtils
-import org.springframework.test.web.reactive.server.WebTestClient
 import uk.gov.justice.digital.hmpps.indexer.integration.IntegrationTest
 import uk.gov.justice.digital.hmpps.indexer.integration.wiremock.CommunityApiExtension
 import uk.gov.justice.digital.hmpps.indexer.integration.wiremock.OAuthExtension
@@ -21,10 +20,6 @@ class HealthCheckIntegrationTest : IntegrationTest() {
 
   @Autowired
   private lateinit var indexQueueHealth: QueueHealth
-
-  @Suppress("SpringJavaInjectionPointsAutowiringInspection")
-  @Autowired
-  private lateinit var webTestClient: WebTestClient
 
   @Autowired
   @Value("\${event.sqs.queue.name}")
