--- src/test/kotlin/uk/gov/justice/digital/hmpps/indexer/integration/ResourceIntegrationTest.kt
@@ -1,32 +1,0 @@
-package uk.gov.justice.digital.hmpps.indexer.integration
-
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.boot.test.mock.mockito.SpyBean
-import org.springframework.http.HttpHeaders
-import org.springframework.test.context.ActiveProfiles
-import org.springframework.test.web.reactive.server.WebTestClient
-import uk.gov.justice.digital.hmpps.indexer.service.IndexService
-import java.time.Duration
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles(profiles = ["test"])
-class ResourceIntegrationTest {
-
-  @Autowired
-  lateinit var webTestClient: WebTestClient
-
-  @Autowired
-  protected lateinit var jwtAuthHelper: JwtAuthHelper
-
-  @SpyBean
-  protected lateinit var indexService: IndexService
-
-  internal fun setAuthorisation(user: String = "probation-offender-search-indexer-client", roles: List<String> = listOf()): (HttpHeaders) -> Unit {
-    val token = jwtAuthHelper.createJwt(subject = user,
-        scope = listOf("read"),
-        expiryTime = Duration.ofHours(1L),
-        roles = roles)
-    return { it.set(HttpHeaders.AUTHORIZATION, "Bearer $token") }
-  }
-}
