+++ src/test/kotlin/uk/gov/justice/digital/hmpps/indexer/service/EntityHelper.kt
@@ -3,6 +3,6 @@ package uk.gov.justice.digital.hmpps.indexer.service
 import com.google.gson.Gson
 
 
-fun anOffender(crn: String = "X12345", offenderId: Long = 99) = Offender(Gson().toJson(OffenderDetail(crn, offenderId)))
+fun anOffender(crn: String = "X12345", offenderId: Long = 99) = Offender(Gson().toJson(OffenderDetail(offenderId, IDs(crn = crn))))
